Make merge propagate props back to both replicas
This fixes an issue where the utime of the conflicting files is kept
at the original utime in each replica even though a successful merge
has occurred. Moreover, the archives are correctly updated to the utime
of the merged file (workingarch), which results in a spurious pending
props change on at least one of the replicas to restore the incorrect
utime (the archive has the correct utime, so the change in utime looks
like a legitimate update).

As a side effect, permissions are also propagated from whatever end up
being the working{1,2} files back to their respective replicas. This
is *not* what we want in general. To resolve this, copyBack should
define newprops as propsTo (as before), but update not only the length
(as before), but also the utime. Unfortunately, the Props and Fileinfo
interfaces don't allow this easily right now and I didn't want to make
major changes.